+++ lib/controller/task_controller.dart
@@ -4,6 +4,7 @@ import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:monitor_app/controller/app_provider.dart';
+import 'package:monitor_app/model/asset.dart';
 import 'package:monitor_app/mstate/task_state.dart';
 
 final taskControllerProvider =
@@ -76,10 +77,30 @@ class TaskController extends AutoDisposeNotifier<TaskState> {
         debugPrint('local task : (${task.id})');
       }
 
-      state = TaskLoaded(tasks: localTasks);
+      state = TasksLoaded(tasks: localTasks);
     } on Exception catch (e) {
       state = TaskLoadedWithError(message: e.toString());
     }
     //
+  }
+
+  updateAssetLocalTask(Asset asset) async {
+    state = TaskLoading();
+    await ref.read(localdataServiceProvider).updateAsset(asset);
+    var localTasks = await ref.read(localdataServiceProvider).getAllTasks();
+    state = TasksLoaded(tasks: localTasks!);
+  }
+
+  updateAssetsLocalTask(List<Asset> assets) async {
+    state = TaskLoading();
+    await ref.read(localdataServiceProvider).updateAssets(assets);
+    var localTasks = await ref.read(localdataServiceProvider).getAllTasks();
+    state = TasksLoaded(tasks: localTasks!);
+  }
+
+  getTaskById(int taskId) async {
+    state = TaskLoading();
+    var task = await ref.read(localdataServiceProvider).getTaskById(taskId);
+    state = TasksLoaded(tasks: [task!]);
   }
 }