+++ lib/router.dart
@@ -1,4 +1,3 @@
-
 import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
 import 'package:monitor_app/model/category_checklist_preventive.dart';
@@ -47,10 +46,11 @@ Route<dynamic>? onGenerateRoute(RouteSettings settings) {
         settings: RouteSettings(name: HomeScreen.routeName),
       );
     case 'task':
-      Task task = settings.arguments as Task;
+      // Task task = settings.arguments as Task;
+      int taskId = settings.arguments as int;
       return PageRouteBuilder(
         pageBuilder: (context, animation, secondaryAnimation) =>
-            TaskScreen(task: task),
+            TaskScreen(taskId: taskId),
         transitionsBuilder: (context, animation, secondaryAnimation, child) {
           const begin = Offset(1.0, 0.0);
           const end = Offset.zero;